+++ crates/lang_srv/src/analysis/completion.rs
@@ -1,255 +1,23 @@
 use std::{collections::HashMap, sync::Arc};
 
-use log::{debug, trace, warn};
+use log::{debug, warn};
 
-use roc_can::{
-    def::Def,
-    expr::{ClosureData, Declarations, Expr, WhenBranch},
-    pattern::{ListPatterns, Pattern, RecordDestruct, TupleDestruct},
-    traverse::{walk_decl, walk_def, walk_expr, DeclarationInfo, Visitor},
-};
+use roc_can::{expr::Declarations, traverse::Visitor};
 use roc_collections::{MutMap, VecMap};
 use roc_load::docs::{DocDef, ModuleDocumentation};
 use roc_module::symbol::{Interns, ModuleId, Symbol};
-use roc_region::all::{Loc, Position, Region};
+use roc_region::all::Position;
 use roc_types::{
     subs::{Subs, Variable},
     types::Alias,
 };
 use tower_lsp::lsp_types::{self, CompletionItem, CompletionItemKind};
 
+use self::visitor::CompletionVisitor;
+
 use super::{utils::format_var_type, ModulesInfo};
 mod formatting;
-
-pub struct CompletionVisitor<'a> {
-    position: Position,
-    found_decls: Vec<(Symbol, Variable)>,
-    pub interns: &'a Interns,
-    pub prefix: String,
-}
-
-impl Visitor for CompletionVisitor<'_> {
-    fn should_visit(&mut self, region: Region) -> bool {
-        region.contains_pos(self.position)
-    }
-
-    fn visit_expr(&mut self, expr: &Expr, region: Region, var: Variable) {
-        if region.contains_pos(self.position) {
-            let mut res = self.expression_defs(expr);
-            self.found_decls.append(&mut res);
-
-            walk_expr(self, expr, var);
-        }
-    }
-
-    fn visit_decl(&mut self, decl: DeclarationInfo<'_>) {
-        match decl {
-            DeclarationInfo::Value { loc_expr, .. }
-            | DeclarationInfo::Function {
-                loc_body: loc_expr, ..
-            }
-            | DeclarationInfo::Destructure { loc_expr, .. } => {
-                let res = self.decl_to_completion_item(&decl);
-                self.found_decls.extend(res);
-                if loc_expr.region.contains_pos(self.position) {
-                    walk_decl(self, decl);
-                };
-            }
-            _ => {
-                walk_decl(self, decl);
-            }
-        }
-    }
-
-    fn visit_def(&mut self, def: &Def) {
-        let res = self.extract_defs(def);
-        self.found_decls.extend(res);
-        walk_def(self, def);
-    }
-}
-impl CompletionVisitor<'_> {
-    fn extract_defs(&mut self, def: &Def) -> Vec<(Symbol, Variable)> {
-        trace!("Completion begin");
-        def.pattern_vars
-            .iter()
-            .map(|(symbol, var)| (*symbol, *var))
-            .collect()
-    }
-    fn expression_defs(&self, expr: &Expr) -> Vec<(Symbol, Variable)> {
-        match expr {
-            Expr::When {
-                expr_var, branches, ..
-            } => self.when_is_expr(branches, expr_var),
-            Expr::Closure(ClosureData {
-                arguments,
-                loc_body,
-                ..
-            }) => {
-                //if we are inside the closure complete it's vars
-                if loc_body.region.contains_pos(self.position) {
-                    arguments
-                        .iter()
-                        .flat_map(|(var, _, pat)| self.patterns(&pat.value, var))
-                        .collect()
-                } else {
-                    vec![]
-                }
-            }
-            _ => vec![],
-        }
-    }
-
-    ///Extract any variables made available by the branch of a when_is expression that contains `self.position`
-    fn when_is_expr(
-        &self,
-        branches: &[WhenBranch],
-        expr_var: &Variable,
-    ) -> Vec<(Symbol, Variable)> {
-        branches
-            .iter()
-            .flat_map(
-                |WhenBranch {
-                     patterns, value, ..
-                 }| {
-                    if value.region.contains_pos(self.position) {
-                        patterns
-                            .iter()
-                            .flat_map(|pattern| self.patterns(&pattern.pattern.value, expr_var))
-                            .collect()
-                    } else {
-                        vec![]
-                    }
-                },
-            )
-            .collect()
-    }
-
-    fn record_destructure(&self, destructs: &[Loc<RecordDestruct>]) -> Vec<(Symbol, Variable)> {
-        destructs
-            .iter()
-            .flat_map(|a| match &a.value.typ {
-                roc_can::pattern::DestructType::Required
-                | roc_can::pattern::DestructType::Optional(_, _) => {
-                    vec![(a.value.symbol, a.value.var)]
-                }
-                roc_can::pattern::DestructType::Guard(var, pat) => self.patterns(&pat.value, var),
-            })
-            .collect()
-    }
-
-    fn tuple_destructure(&self, destructs: &[Loc<TupleDestruct>]) -> Vec<(Symbol, Variable)> {
-        destructs
-            .iter()
-            .flat_map(|a| {
-                let (var, pattern) = &a.value.typ;
-                self.patterns(&pattern.value, var)
-            })
-            .collect()
-    }
-
-    fn list_pattern(&self, list_elems: &ListPatterns, var: &Variable) -> Vec<(Symbol, Variable)> {
-        list_elems
-            .patterns
-            .iter()
-            .flat_map(|a| self.patterns(&a.value, var))
-            .collect()
-    }
-    fn tag_pattern(&self, arguments: &[(Variable, Loc<Pattern>)]) -> Vec<(Symbol, Variable)> {
-        arguments
-            .iter()
-            .flat_map(|(var, pat)| self.patterns(&pat.value, var))
-            .collect()
-    }
-
-    fn as_pattern(
-        &self,
-        as_pat: &Pattern,
-        as_symbol: Symbol,
-        var: &Variable,
-    ) -> Vec<(Symbol, Variable)> {
-        //Get the variables introduced within the pattern
-        let mut patterns = self.patterns(as_pat, var);
-        //Add the "as" that wraps the whole pattern
-        patterns.push((as_symbol, *var));
-        patterns
-    }
-    ///Returns a list of symbols defined by this pattern.  
-    ///`pattern_var`: Variable type of the entire pattern. This will be returned if the pattern turns out to be an identifier
-    fn patterns(
-        &self,
-        pattern: &roc_can::pattern::Pattern,
-        pattern_var: &Variable,
-    ) -> Vec<(Symbol, Variable)> {
-        match pattern {
-            roc_can::pattern::Pattern::Identifier(symbol) => {
-                if self.is_match(symbol) {
-                    vec![(*symbol, *pattern_var)]
-                } else {
-                    vec![]
-                }
-            }
-            Pattern::AppliedTag { arguments, .. } => self.tag_pattern(arguments),
-            Pattern::UnwrappedOpaque { argument, .. } => {
-                self.patterns(&argument.1.value, &argument.0)
-            }
-            Pattern::List {
-                elem_var, patterns, ..
-            } => self.list_pattern(patterns, elem_var),
-            roc_can::pattern::Pattern::As(pat, symbol) => {
-                self.as_pattern(&pat.value, *symbol, pattern_var)
-            }
-            roc_can::pattern::Pattern::RecordDestructure { destructs, .. } => {
-                self.record_destructure(destructs)
-            }
-            roc_can::pattern::Pattern::TupleDestructure { destructs, .. } => {
-                self.tuple_destructure(destructs)
-            }
-            _ => vec![],
-        }
-    }
-
-    fn is_match(&self, symbol: &Symbol) -> bool {
-        symbol.as_str(self.interns).starts_with(&self.prefix)
-    }
-
-    fn decl_to_completion_item(&self, decl: &DeclarationInfo) -> Vec<(Symbol, Variable)> {
-        match decl {
-            DeclarationInfo::Value {
-                expr_var, pattern, ..
-            } => self.patterns(pattern, expr_var),
-            DeclarationInfo::Function {
-                expr_var,
-                pattern,
-                function,
-                loc_body,
-                ..
-            } => {
-                let mut out = vec![];
-                //Append the function declaration itself for recursive calls
-                out.extend(self.patterns(pattern, expr_var));
-
-                if loc_body.region.contains_pos(self.position) {
-                    //also add the arguments if we are inside the function
-                    let args = function
-                        .value
-                        .arguments
-                        .iter()
-                        .flat_map(|(var, _, pat)| self.patterns(&pat.value, var));
-                    //We add in the pattern for the function declaration
-                    out.extend(args);
-                    trace!("Added function args to completion output =:{:#?}", out);
-                }
-                out
-            }
-            DeclarationInfo::Destructure {
-                loc_pattern,
-                expr_var,
-                ..
-            } => self.patterns(&loc_pattern.value, expr_var),
-            DeclarationInfo::Expectation { .. } => vec![],
-        }
-    }
-}
+mod visitor;
 
 fn get_completions(
     position: Position,
@@ -267,45 +35,6 @@ fn get_completions(
     visitor.found_decls
 }
 
-fn make_completion_item(
-    subs: &mut Subs,
-    module_id: &ModuleId,
-    interns: &Interns,
-    docs: Option<String>,
-    str: String,
-    var: Variable,
-) -> CompletionItem {
-    let type_str = format_var_type(var, subs, module_id, interns);
-    let typ = match subs.get(var).content {
-        roc_types::subs::Content::Structure(var) => match var {
-            roc_types::subs::FlatType::Apply(_, _) => CompletionItemKind::FUNCTION,
-            roc_types::subs::FlatType::Func(_, _, _) => CompletionItemKind::FUNCTION,
-            roc_types::subs::FlatType::EmptyTagUnion
-            | roc_types::subs::FlatType::TagUnion(_, _) => CompletionItemKind::ENUM,
-            _ => CompletionItemKind::VARIABLE,
-        },
-        a => {
-            debug!(
-                "No specific completionKind for variable type: {:?} defaulting to 'Variable'",
-                a
-            );
-            CompletionItemKind::VARIABLE
-        }
-    };
-
-    CompletionItem {
-        label: str,
-        detail: Some(type_str),
-        kind: Some(typ),
-        documentation: docs.map(|d| {
-            lsp_types::Documentation::MarkupContent(lsp_types::MarkupContent {
-                kind: lsp_types::MarkupKind::Markdown,
-                value: d,
-            })
-        }),
-        ..Default::default()
-    }
-}
 /// Walks through declarations that would be accessible from the provided position adding them to a list of completion items until all accessible declarations have been fully explored
 pub fn get_completion_items(
     position: Position,
@@ -322,6 +51,7 @@ pub fn get_completion_items(
     debug!("extended with:{:#?}", exposed_imports);
     make_completion_items(subs, module_id, interns, docs, completions)
 }
+
 pub(super) fn get_module_completion_items(
     prefix: String,
     interns: &Interns,
@@ -390,7 +120,7 @@ fn get_module_exposed_completion(
                         subs,
                         mod_id,
                         interns,
-                        completion_docs.remove(&sym),
+                        completion_docs.remove(sym),
                         sym.as_str(interns).to_string(),
                         *var,
                     )
@@ -398,6 +128,26 @@ fn get_module_exposed_completion(
                 .expect("Couldn't find subs for module during completion.")
         })
         .collect::<Vec<_>>()
+}
+
+///Efficently walks the list of docs collecting the docs for  completions as we go. Should be faster than re-walking for every completion
+fn get_completion_docs(
+    completions: &[(Symbol, Variable)],
+    docs: &ModuleDocumentation,
+) -> HashMap<Symbol, String> {
+    let mut symbols = completions.iter().map(|(s, _)| s).collect::<Vec<_>>();
+    docs.entries
+        .iter()
+        .filter_map(|doc| match doc {
+            roc_load::docs::DocEntry::DocDef(DocDef { docs, symbol, .. }) => {
+                let docs = docs.clone()?;
+                let (idx, _s) = symbols.iter().enumerate().find(|(_i, s)| s == &&symbol)?;
+                symbols.swap_remove(idx);
+                Some((*symbol, docs))
+            }
+            _ => None,
+        })
+        .collect()
 }
 
 ///Provides a list of completions for Type aliases within the scope.
@@ -421,6 +171,7 @@ fn _alias_completions(
         })
         .collect()
 }
+
 fn make_completion_items(
     subs: &mut Subs,
     module_id: &ModuleId,
@@ -446,6 +197,7 @@ fn make_completion_items(
         })
         .collect()
 }
+
 fn make_completion_items_string(
     subs: &mut Subs,
     module_id: &ModuleId,
@@ -457,24 +209,52 @@ fn make_completion_items_string(
         .map(|(symbol, var)| make_completion_item(subs, module_id, interns, None, symbol, var))
         .collect()
 }
-///Efficently walks the list of docs collecting the docs for  completions as we go. Should be faster than re-walking for every completion
-fn get_completion_docs(
-    completions: &[(Symbol, Variable)],
-    docs: &ModuleDocumentation,
-) -> HashMap<Symbol, String> {
-    let mut symbols = completions.iter().map(|(s, _)| s).collect::<Vec<_>>();
-    docs.entries
-        .iter()
-        .filter_map(|doc| match doc {
-            roc_load::docs::DocEntry::DocDef(DocDef { docs, symbol, .. }) => {
-                let docs = docs.clone()?;
-                let (idx, s) = symbols.iter().enumerate().find(|(i, s)| s == &&symbol)?;
-                symbols.swap_remove(idx);
-                Some((*symbol, docs))
-            }
-            _ => None,
-        })
-        .collect()
+
+fn make_completion_item(
+    subs: &mut Subs,
+    module_id: &ModuleId,
+    interns: &Interns,
+    docs: Option<String>,
+    str: String,
+    var: Variable,
+) -> CompletionItem {
+    let type_str = format_var_type(var, subs, module_id, interns);
+    let typ = match subs.get(var).content {
+        roc_types::subs::Content::Structure(var) => match var {
+            roc_types::subs::FlatType::Apply(_, _) => CompletionItemKind::FUNCTION,
+            roc_types::subs::FlatType::Func(_, _, _) => CompletionItemKind::FUNCTION,
+            roc_types::subs::FlatType::EmptyTagUnion
+            | roc_types::subs::FlatType::TagUnion(_, _) => CompletionItemKind::ENUM,
+            _ => CompletionItemKind::VARIABLE,
+        },
+        a => {
+            debug!(
+                "No specific completionKind for variable type: {:?} defaulting to 'Variable'",
+                a
+            );
+            CompletionItemKind::VARIABLE
+        }
+    };
+
+    CompletionItem {
+        label: str,
+        detail: Some(type_str),
+        kind: Some(typ),
+        documentation: docs.map(|d| {
+            lsp_types::Documentation::MarkupContent(lsp_types::MarkupContent {
+                kind: lsp_types::MarkupKind::Markdown,
+                value: d,
+            })
+        }),
+        ..Default::default()
+    }
+}
+
+struct FieldCompletion {
+    ///The name of the variable that is a record
+    variable_name: String,
+    field: String,
+    middle_fields: Vec<String>,
 }
 
 ///Finds the types of and names of all the fields of a record
@@ -532,12 +312,6 @@ fn find_record_fields(var: Variable, subs: &mut Subs) -> Vec<(String, Variable)>
     }
 }
 
-struct FieldCompletion {
-    ///The name of the variable that is a record
-    variable_name: String,
-    field: String,
-    middle_fields: Vec<String>,
-}
 ///Splits a completion prefix for a field into its components
 ///E.g. a.b.c.d->{variable_name:"a",middle_fields:["b","c"],field:"d"}
 fn get_field_completion_parts(symbol_prefix: &str) -> Option<FieldCompletion> {
